+++ test/neon-test-app/app/components/BasicCount.vue
@@ -23,7 +23,7 @@ const result = ref(-2) // initial state for test
 async function doCount() {
   result.value = await count({
     from: 'playing_with_neon',
-    where: [{ column: 'name', condition: 'LIKE', value: '\'test%\'' }],
+    where: [{ column: 'name', operator: 'LIKE', value: '\'test%\'' }],
   }) as number
 }
 </script>