+++ test/neon-test-app/app/components/BasicUpdate.vue
@@ -24,7 +24,7 @@ async function doUpdate() {
   result.value = await update({
     table: 'playing_with_neon',
     values: { value: '1' },
-    where: [{ column: 'name', condition: '=', value: '\'test\'' }],
+    where: [{ column: 'name', operator: '=', value: '\'test\'' }],
   }) as string
 }
 </script>