+++ test/neon-test-app/app/pages/TestDeleteSchemaAlias.vue
@@ -36,7 +36,7 @@ const result = ref('')
 async function doDelete() {
   result.value = await del({
     table: { schema: 'neon2', table: 'playing_with_neon', alias: 'p1' },
-    where: [{ column: { alias: 'p1', name: 'id' }, condition: '>', value: '10' }],
+    where: [{ column: { alias: 'p1', name: 'id' }, operator: '>', value: '10' }],
   }) as string
 }
 </script>